UDP.write[v]: make src_port mandatory, allow optional src
if [src] is specified, set that address to be the sender IP in the
packet.  If not, use [t.src].  Also, don't allow the caller to ask us to
set a random source port; the only implementation around doesn't know
how to do it, which will result in a runtime failure.