make argv_qubes, which passes only valid runtime_keys to bootvar
This will only pass those runtime arguments which are expected to
the Functoria runtime, and therefore to Cmdliner.  Passing unexpected
arguments to Functoria results in runtime crashes, so this is needed
to reliably boot in environments where the supplied arguments can't be
easily controlled, such as Qubes or an external cloud hosting service.

This needs a version of mirage-bootvar-xen that understands
`Bootvar.argv ~filter`, and a version of Functoria that exposes
`Functoria_key.name` and writes a list of these values to `key_gen.ml`.
PR #23 in https://github.com/mirage/mirage-bootvar-xen provides the filter,
and PR #68 to https://github.com/mirage/functoria provides the runtime arguments.