test unikernel that uses a lot of impls
test/config.ml and test/unikernel.ml represent an attempt to use the
largest possible set of functions that yield a 'a impl from mirage.mli .
Some will only be available when configured for certain modes.  The
unikernel itself is a noop (and will be difficult to run under most
circumstances, as it wants a lot of network and block devices), but
successfully compiling implies that the packages referenced and the code
generated by `mirage configure` are in syntactic agreement on the
interfaces used.